Autorizacion en los productos

+++ src/Controllers/product.js
@@ -5,11 +5,17 @@ const autorization = require('../Middlewares/autorization');
 const route = express.Router();
 
 //create users
-route.post("/products", (req,res) => {
+route.post("/products",autorization, (req,res) => {
     const product = productSchema(req.body);
-    product.save()
-        .then((data) => res.json(data))
-        .catch((error) => res.json({ message: error }));
+    const { userRol } = req;
+    
+    if (userRol == 'admin') {
+        product.save()
+            .then((data) => res.json(data))
+            .catch((error) => res.json({ message: error }));
+    } else {
+        res.status(401).json({ error: "You do not have the permissions to perform this action with the credentials provided." })
+    }
 
 });
 
@@ -51,12 +57,18 @@ route.put("/products/:id", autorization, (req, res) => {
 });
 
 //delete product
-route.delete("/products/:id", (req, res) => {
+route.delete("/products/:id",autorization, (req, res) => {
     const { id } = req.params;
-    productSchema
-        .deleteOne({ _id: id })
-        .then((data) => res.json(data))
-        .catch((error) => res.json({ message: error }));
+    const { userRol } = req;
+    
+    if (userRol == 'admin') {
+        productSchema
+            .deleteOne({ _id: id })
+            .then((data) => res.json(data))
+            .catch((error) => res.json({ message: error }));
+    } else {
+        res.status(401).json({ error: "You do not have the permissions to perform this action with the credentials provided." })
+    }
 
 });
